+++ react/features/mobile/navigation/functions.tsx
@@ -1,44 +1,14 @@
-import React, { useCallback } from 'react';
-import { useTranslation } from 'react-i18next';
-import { GestureResponderEvent, Platform } from 'react-native';
-import { useDispatch } from 'react-redux';
+import React from 'react';
+import { GestureResponderEvent } from 'react-native';
 
-
-import { appNavigate } from '../../app/actions.native';
 import { IStateful } from '../../base/app/types';
 import { PREJOIN_PAGE_ENABLED } from '../../base/flags/constants';
 import { getFeatureFlag } from '../../base/flags/functions';
-import { IconCloseLarge } from '../../base/icons/svg';
 import { toState } from '../../base/redux/functions';
-import { cancelKnocking } from '../../lobby/actions.native';
 import { isPrejoinEnabledInConfig } from '../../prejoin/functions.native';
 
-import HeaderNavigationButton from './components/HeaderNavigationButton';
-
-/**
- * Screen close button component.
- */
-const ScreenCloseButton = React.memo(({ goBack, isIos }: { goBack: (e?: GestureResponderEvent | React.MouseEvent) => void; isIos: boolean }) => {
-    const { t } = useTranslation();
-
-    if (isIos) {
-        return (
-            <HeaderNavigationButton
-                id = { 'close-screen-button' }
-                label = { t('dialog.close') }
-                onPress = { goBack } />
-        );
-    }
-
-    return (
-        <HeaderNavigationButton
-            id = { 'close-screen-button' }
-            onPress = { goBack }
-            src = { IconCloseLarge } />
-    );
-});
-
-ScreenCloseButton.displayName = 'ScreenCloseButton';
+import LobbyCloseButton from './components/LobbyCloseButton';
+import ScreenCloseButton from './components/ScreenCloseButton';
 
 /**
  * Close icon/text button based on platform.
@@ -47,7 +17,7 @@ ScreenCloseButton.displayName = 'ScreenCloseButton';
  * @returns {React.Component}
  */
 export function screenHeaderCloseButton(goBack: (e?: GestureResponderEvent | React.MouseEvent) => void) {
-    return () => <ScreenCloseButton goBack = { goBack } isIos = { Platform.OS === 'ios' } />;
+    return () => <ScreenCloseButton goBack = { goBack } />;
 }
 
 /**
@@ -66,40 +36,10 @@ export function isPrejoinPageEnabled(stateful: IStateful) {
 }
 
 /**
- * Lobby close button component.
- */
-const LobbyCloseButton = React.memo(() => {
-    const dispatch = useDispatch();
-    const { t } = useTranslation();
-    const goBack = useCallback(() => {
-        dispatch(cancelKnocking());
-        dispatch(appNavigate(undefined));
-    }, [ dispatch ]);
-
-    if (Platform.OS === 'ios') {
-        return (
-            <HeaderNavigationButton
-                id = { 'close-screen-button' }
-                label = { t('dialog.close') }
-                onPress = { goBack } />
-        );
-    }
-
-    return (
-        <HeaderNavigationButton
-            id = { 'close-screen-button' }
-            onPress = { goBack }
-            src = { IconCloseLarge } />
-    );
-});
-
-LobbyCloseButton.displayName = 'LobbyCloseButton';
-
-/**
  * Close icon/text button for lobby screen based on platform.
  *
  * @returns {React.Component}
  */
 export function lobbyScreenHeaderCloseButton() {
-    return () => <LobbyCloseButton />;
+    return LobbyCloseButton;
 }
